upgrade logging-operator install
